@@ -35,8 +35,8 @@
     </build>
 
     <properties>
-        <maven.compiler.source>12</maven.compiler.source>
-        <maven.compiler.target>12</maven.compiler.target>
+        <maven.compiler.source>24</maven.compiler.source>
+        <maven.compiler.target>24</maven.compiler.target>
         <vavr.version>0.10.2</vavr.version>
         <assertj-core.version>3.23.1</assertj-core.version>
         <commons-lang3.version>3.12.0</commons-lang3.version>
