@@ -1,13 +1,14 @@
 package com.baeldung.libraries.h2;
 
-import org.junit.jupiter.api.Assertions;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import org.junit.jupiter.api.Test;
 import java.sql.*;
 
 public class H2JUnitTest {
 
     @Test
-    public static void whenConnectingToSpecificSchema_ShouldReturnTheSpecificSchema() throws Exception {
+    public void whenConnectingToSpecificSchema_ShouldReturnTheSpecificSchema() throws Exception {
         Connection conn = DriverManager.getConnection(
           "jdbc:h2:~/TEST_DB;INIT=CREATE SCHEMA IF NOT EXISTS TEST_SCHEMA\\;SET SCHEMA TEST_SCHEMA",
           "sa",
@@ -21,7 +22,7 @@ public class H2JUnitTest {
         String actualSchema = rs.getString(1);
         String expectedSchema = new String("TEST_SCHEMA");
 
-        Assertions.assertTrue(actualSchema.equals(expectedSchema));
+        assertTrue(actualSchema.equals(expectedSchema));
     }
 
 }
