@@ -1,10 +1,13 @@
 package com.baeldung.resttemplate.web.service;
 
+import com.baeldung.resttemplate.web.exception.UnauthorizedException;
 import com.baeldung.resttemplate.web.handler.RestTemplateResponseErrorHandler;
 import com.baeldung.resttemplate.web.model.Bar;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.client.RestTemplate;
 
 @Service
@@ -15,12 +18,20 @@ public class BarConsumerService {
     @Autowired
     public BarConsumerService(RestTemplateBuilder restTemplateBuilder) {
         restTemplate = restTemplateBuilder
-          .errorHandler(new RestTemplateResponseErrorHandler())
-          .build();
+                .errorHandler(new RestTemplateResponseErrorHandler())
+                .build();
     }
 
     public Bar fetchBarById(String barId) {
-        return restTemplate.getForObject("/bars/4242", Bar.class);
+        try {
+            return restTemplate.getForObject("/bars/" + barId, Bar.class);
+        } catch (HttpStatusCodeException e) {
+            if (HttpStatus.UNAUTHORIZED == e.getStatusCode()) {
+                String responseBody = e.getResponseBodyAsString();
+                throw new UnauthorizedException("Unauthorized access: " + responseBody);
+            }
+            throw e;
+        }
     }
 
 }
